fix: Changed price to Float Field

+++ flightly/flight_booking/models.py
@@ -32,11 +32,9 @@ class Flight(models.Model):
         _("Arrival Airport"), max_length=50, null=False)
     departure_datetime = models.DateTimeField(_("Departure Time"))
     capacity = models.IntegerField(_("Carrying Capacity"), null=False)
-    price = models.DecimalField(
+    price = models.FloatField(
         _("Ticket Price"),
-        decimal_places=2,
-        max_digits=10,
-        default=0)
+        default=0.00)
 
     @property
     def available_seats(self):
